+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## 1.1.7 (Jan 28, 2022)
+## 1.1.7 (Jan 31, 2022)
 * Fix test code.
 * Fix regression bug: worker-pool did not respect `worker-count` for async jobs.
 * Fix bug: an exception from (n)ack blocks whole consumer.
+* Perf: A slow worker does not block `com.climate.squeedo.sqs-consumer/acker`.
 
 ## 1.1.6 (Jan 24, 2022)
 * **BREAKING**